Reference

RBMsAnnealedImportanceSampling.ais — Method

ais(rbm0, rbm1, v0, βs)

Provided v0 is an equilibrated sample from rbm0, returns F such that mean(exp.(F)) is an unbiased estimator of Z1/Z0, the ratio of partition functions of rbm1 and rbm0.

!!! tip Use logmeanexp logmeanexp(F), using the function logmeanexp[@ref] provided in this package, tends to give a better approximation of log(Z1) - log(Z0) than mean(F).

RBMsAnnealedImportanceSampling.aise — Method

aise(rbm, [βs]; [nbetas], init=rbm.visible, nsamples=1)

AIS estimator of the log-partition function of rbm. It is recommended to fit init to the single-site statistics of rbm (or the data).

!!! tip Use large nbetas For more accurate estimates, use larger nbetas. It is usually better to have large nbetas and small nsamples, rather than large nsamples and small nbetas.

RBMsAnnealedImportanceSampling.anneal — Method

anneal(rbm0, rbm1; β)

Returns an RBM that interpolates between rbm0 and rbm1. Denoting by E0(v, h) and E1(v, h) the energies assigned by rbm0 and rbm1, respectively, the returned RBM assigns energies given by:

E(v,h) = (1 - β) * E0(v) + β * E1(v, h)

RBMsAnnealedImportanceSampling.log_partition_zero_weight — Method

log_partition_zero_weight(rbm)

Log-partition function of a zero-weight version of rbm.

RBMsAnnealedImportanceSampling.logmeanexp — Method

logmeanexp(A; dims=:)

Computes log.(mean(exp.(A); dims)), in a numerically stable way.

RBMsAnnealedImportanceSampling.logstdexp — Method

logstdexp(A; dims=:)

Computes log.(std(exp.(A); dims)), in a numerically stable way.

RBMsAnnealedImportanceSampling.logvarexp — Method

logvarexp(A; dims=:)

Computes log.(var(exp.(A); dims)), in a numerically stable way.

RBMsAnnealedImportanceSampling.raise — Method

raise(rbm::RBM, βs; v, init)

Reverse AIS estimator of the log-partition function of rbm. While aise tends to understimate the log of the partition function, raise tends to overestimate it. v must be an equilibrated sample from rbm.

!!! tip Use logmeanexp If F = raise(...), then -logmeanexp(-F), using the function logmeanexp[@ref] provided in this package, tends to give a better approximation of log(Z) than mean(F).

!!! tip Sandwiching the log-partition function If Rf = aise(...), Rr = raise(...) are the AIS and reverse AIS estimators, we have the stochastic bounds logmeanexp(Rf) ≤ log(Z) ≤ -logmeanexp(-Rr).